requirements.txt
amqp==2.5.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-celery-results==1.1.2
django-crispy-forms==1.7.2
django-jet==1.0.8
django-phone-field==1.7.1
django-recaptcha2==1.4.1
django-seed==0.1.9
Faker==1.0.7
gunicorn==19.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.6.1
mysqlclient==1.4.4
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
Pygments==2.4.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
recaptcha==1.0rc1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.1.2
